+++ db/migrate/20241015220323_create_transactions.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class CreateTransactions < ActiveRecord::Migration[7.0]
   def change
     create_enum :transaction_type, %w[buy sell]
@@ -5,11 +7,11 @@ class CreateTransactions < ActiveRecord::Migration[7.0]
     create_table :transactions do |t|
       t.references :user, null: false, foreign_key: true
       t.enum :current_transaction, enum_type: :transaction_type
-      t.decimal :amount_sent
+      t.float :amount_sent
       t.string :currency_sent
-      t.decimal :amount_received
+      t.float :amount_received
       t.string :currency_received
-      t.decimal :exchange_rate, precision: 20, scale: 8
+      t.float :exchange_rate, precision: 20, scale: 8
 
       t.timestamps
     end
